requirements.txt
python-telegram-bot==12.4.2
requests=2.22.0
bs4=4.8.0